+++ src/main/kotlin/cn/inrhor/questengine/command/main/TagsCommand.kt
@@ -1,0 +1,19 @@
+package cn.inrhor.questengine.command.main
+
+import cn.inrhor.questengine.command.tags.TagsAdd
+import cn.inrhor.questengine.command.tags.TagsList
+import cn.inrhor.questengine.command.tags.TagsRemove
+import taboolib.common.platform.command.*
+
+object TagsCommand {
+
+    @CommandBody
+    val list = TagsList.list
+
+    @CommandBody
+    val add = TagsAdd.add
+
+    @CommandBody
+    val remove = TagsRemove.remove
+
+}
